Added Filter interface
